Update cemu.po
“体感加速器” means "Motion detection accelerator", this doesn't make sense. MotionPlus does not "accelerate" motion detection, but "improve" or "promote" or "boost" the accuracy of motion detection.

(According to Wikipedia : The Wii MotionPlus is an expansion device for the Wii Remote video game controller for the Wii that allows it to capture complex motion more accurately. According to Nintendo, the sensor in the device supplements the accelerometer and Sensor Bar capabilities of the Wii Remote to enable actions to be rendered identically on the screen in real time.)

If it has to be translated into Chinese, I suggest "体感增强器", which means "Motion detection promoter" or "Motion detection booster".

And, if search "体感增强器" on Baidu (China's Google), we can get a bunch of useful results about Nintendo Wii MotionPlus.